+++ src/main/java/io/github/astrapi69/jdbc/postgresql/PostgreSQLConnectionsExtensions.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.jdbc;
+package io.github.astrapi69.jdbc.postgresql;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -32,6 +32,9 @@ import java.util.Properties;
 
 import lombok.NonNull;
 import lombok.experimental.UtilityClass;
+import io.github.astrapi69.jdbc.CreationState;
+import io.github.astrapi69.jdbc.JdbcConnectionInfo;
+import io.github.astrapi69.jdbc.JdbcUrlBean;
 
 /**
  * The class {@link PostgreSQLConnectionsExtensions} have convenience methods to create and connect
@@ -51,10 +54,16 @@ public final class PostgreSQLConnectionsExtensions
 	public static final String URL_PREFIX = "jdbc:postgresql://";
 
 	/** Constant for the port where the PostgreSQL-database listen. */
-	public static final int POSTGRESQL_PORT = 5432;
+	public static final int DEFAULT_POSTGRESQL_PORT = 5432;
 
 	/** Constant for the default user from PostgreSQL-database. */
-	public static final String POSTGRESQL_DEFAULT_USER = "postgres";
+	public static final String DEFAULT_USER = "postgres";
+
+	/** Constant for the default user from PostgreSQL-database. */
+	public static final String DEFAULT_HOST = "localhost";
+
+	/** Constant for the default password from PostgreSQL-database. */
+	public static final String DEFAULT_PASSWORD = "postgres";
 
 	public static final String APP_DB_HOST = "app.db-host";
 	public static final String APP_DB_PORT = "app.db-port";
@@ -118,7 +127,7 @@ public final class PostgreSQLConnectionsExtensions
 		final @NonNull String databaseName, final @NonNull String dbuser,
 		final @NonNull String dbpasswort)
 	{
-		return existsDatabase(hostname, POSTGRESQL_PORT, databaseName, dbuser, dbpasswort);
+		return existsDatabase(hostname, DEFAULT_POSTGRESQL_PORT, databaseName, dbuser, dbpasswort);
 	}
 
 	/**
@@ -173,7 +182,7 @@ public final class PostgreSQLConnectionsExtensions
 		final @NonNull String databaseName, final @NonNull String dbuser,
 		final @NonNull String dbpasswort) throws ClassNotFoundException, SQLException
 	{
-		return getConnection(hostname, POSTGRESQL_PORT, databaseName, dbuser, dbpasswort);
+		return getConnection(hostname, DEFAULT_POSTGRESQL_PORT, databaseName, dbuser, dbpasswort);
 	}
 
 	/**
@@ -229,6 +238,22 @@ public final class PostgreSQLConnectionsExtensions
 		final @NonNull String dbpasswort) throws SQLException
 	{
 		return DriverManager.getConnection(url, dbuser, dbpasswort);
+	}
+
+	/**
+	 * Gets the postgres sql connection with the given url and the default user and the default
+	 * password
+	 *
+	 * @param url
+	 *            the database url
+	 * @return the postgres sql connection
+	 * @throws SQLException
+	 *             is thrown if a database access error occurs or this method is called on a closed
+	 *             connection
+	 */
+	public static Connection getConnection(final @NonNull String url) throws SQLException
+	{
+		return getConnection(url, DEFAULT_USER, DEFAULT_PASSWORD);
 	}
 
 	/**
@@ -296,7 +321,7 @@ public final class PostgreSQLConnectionsExtensions
 		final @NonNull String dbPassword, final @NonNull String characterSet,
 		final @NonNull String collate) throws SQLException, ClassNotFoundException
 	{
-		return newDatabase(hostname, POSTGRESQL_PORT, databaseName, dbUser, dbPassword,
+		return newDatabase(hostname, DEFAULT_POSTGRESQL_PORT, databaseName, dbUser, dbPassword,
 			characterSet, collate);
 	}
 
@@ -406,8 +431,8 @@ public final class PostgreSQLConnectionsExtensions
 		final @NonNull String databaseName, final @NonNull String dbUser,
 		final @NonNull String dbPassword) throws SQLException, ClassNotFoundException
 	{
-		return PostgreSQLConnectionsExtensions.newDatabase(hostname, POSTGRESQL_PORT, databaseName,
-			dbUser, dbPassword);
+		return PostgreSQLConnectionsExtensions.newDatabase(hostname, DEFAULT_POSTGRESQL_PORT,
+			databaseName, dbUser, dbPassword);
 	}
 
 	/**
@@ -428,8 +453,8 @@ public final class PostgreSQLConnectionsExtensions
 	public static CreationState newDatabase(final @NonNull String hostname,
 		final @NonNull String databaseName) throws SQLException, ClassNotFoundException
 	{
-		return PostgreSQLConnectionsExtensions.newDatabase(hostname, databaseName,
-			POSTGRESQL_DEFAULT_USER, POSTGRESQL_DEFAULT_USER);
+		return PostgreSQLConnectionsExtensions.newDatabase(hostname, databaseName, DEFAULT_USER,
+			DEFAULT_USER);
 	}